requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.9
sqlalchemy==1.4.52
databases==0.6.0
httpx==0.27.0
python-dotenv==1.0.0
numpy
pandas
python-multipart
requests
aiohttp
pillow
mysqlclient
pika 
poetry

# Optional for authentication and authorization
passlib[bcrypt]==1.7.4
python-jose==3.3.0

# Optional for handling CORS
# fastapi-cors==0.1.1

# Optional for handling asynchronous tasks
celery==5.2.7
# redis==4.7.1

# Optional for testing
# pytest==7.3.1
# pytest-asyncio==0.21.0
# httpx==0.24.1